requirements.txt
flask==3.0.0
line-bot-sdk==3.5.0
anthropic==0.23.0
google-auth==2.28.0
google-auth-oauthlib==1.2.0
google-api-python-client==2.121.0
requests==2.31.0
pyyaml==6.0.1
python-dotenv==1.0.1
gunicorn==21.2.0
